refactor(portal-e2e): factor out interactions, questions, and tasks
